requirements.txt
cryptography==41.0.3
requests==2.31.0
pytest-cov==4.1.0
